+++ src/constant/router.ts
@@ -1,0 +1,8 @@
+export const ROUTE = {
+  HOME: '/',
+  DETAIL: '/detail',
+  CART: '/cart',
+  ORDER: '/order',
+  ORDER_LIST: '/order-list',
+  ORDER_DETAIL: '/order-detail',
+};
