+++ src/index.tsx
@@ -3,6 +3,7 @@ import ReactDOM from 'react-dom/client';
 import App from './App';
 import './style/css/index.css';
 import { worker } from './mocks/browser';
+import { BrowserRouter } from 'react-router-dom';
 
 // if (process.env.NODE_ENV === 'development') {
 worker.start();
@@ -13,6 +14,8 @@ const root = ReactDOM.createRoot(
 );
 root.render(
   <React.StrictMode>
-    <App />
+    <BrowserRouter>
+      <App />
+    </BrowserRouter>
   </React.StrictMode>
 );
